+++ tests/testthat/test-file_tools.R
@@ -27,7 +27,26 @@ test_that("save_check() does not overwrite incorrectly", {
 })
 
 test_that("dir_...() functions run", {
-  expect_type(dir_contents("../testdata"), "list")
   expect_type(dir_files("../testdata"), "list")
   expect_type(dir_size("../testdata"), "double")
 })
+
+test_that("dir_contents() is correct", {
+
+  x1 <- dir_contents(".", recurse = 1)
+  x2 <- dir_contents(getwd(), recurse = 1)
+  x3 <- dir_contents(paste0(getwd(), "/"), recurse = 1)
+  expect_true(all(x1$level %in% 0:1))
+  expect_equal(x1$level[1], 0L)
+  expect_type(x1, "list")
+  expect_identical(x1, x2, ignore_attr = "call")
+  expect_identical(x1, x3, ignore_attr = "call")
+  x4 <- dir_contents(".", recurse = 1, shorten = FALSE)
+  x5 <- dir_contents(getwd(), recurse = 1, shorten = FALSE)
+  x6 <- dir_contents(paste0(getwd(), "/"), recurse = 1, shorten = FALSE)
+  expect_identical(x1[, -1], x4[, -1], ignore_attr = "call")
+  expect_identical(x4, x5, ignore_attr = "call")
+  expect_identical(x4, x6, ignore_attr = "call")
+})
+
+